+++ drape_frontend/gui/copyright_label.cpp
@@ -74,13 +74,13 @@ drape_ptr<ShapeRenderer> CopyrightLabel::Draw(m2::PointF & size, ref_ptr<dp::Tex
   StaticLabel::CacheStaticText("Map data © OpenStreetMap", "", m_position.m_anchor,
                                DrapeGui::GetGuiTextFont(), tex, result);
 
-  dp::AttributeProvider provider(1 /*stream count*/, result.m_buffer.size());
+  dp::AttributeProvider provider(1 /*stream count*/, static_cast<uint32_t>(result.m_buffer.size()));
   provider.InitStream(0 /*stream index*/, StaticLabel::Vertex::GetBindingInfo(),
                       make_ref(result.m_buffer.data()));
 
-  size_t vertexCount = result.m_buffer.size();
+  uint32_t vertexCount = static_cast<uint32_t>(result.m_buffer.size());
   ASSERT(vertexCount % dp::Batcher::VertexPerQuad == 0, ());
-  size_t indexCount = dp::Batcher::IndexPerQuad * vertexCount / dp::Batcher::VertexPerQuad;
+  uint32_t indexCount = dp::Batcher::IndexPerQuad * vertexCount / dp::Batcher::VertexPerQuad;
 
   size = m2::PointF(result.m_boundRect.SizeX(), result.m_boundRect.SizeY());
   drape_ptr<dp::OverlayHandle> handle = make_unique_dp<CopyrightHandle>(EGuiHandle::GuiHandleCopyright,
